Changed aws methods
I can confirm that the json file of the usermap is properly uploaded
onto my bucket. Currently working on the download method because the
file that is retrieved is in the form of a S3Object and not a File,
gonna ask the prof on what to do with the object and how to convert it
back to a json file and connect it with the other usermap methods that
interact with the usermap stored locally.